refactor callback
