+++ swanlab/data/run/__init__.py
@@ -7,7 +7,7 @@ r"""
 @Description:
     在此处导出SwanLabRun类，一次实验运行应该只有一个SwanLabRun实例
 """
-from .main import SwanLabRun, get_run, except_handler, clean_handler, SwanLabRunState
+from .main import SwanLabRun, get_run, SwanLabRunState
 
 
 def register(*args, **kwargs) -> SwanLabRun:
